+++ CKLDrawing/CKLDrawing.csproj
@@ -20,9 +20,6 @@
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         <PrivateAssets>all</PrivateAssets>
       </PackageReference>
-      <PackageReference Include="MSTest" Version="3.8.3" />
-      <PackageReference Include="MSTest.TestAdapter" Version="3.8.3" />
-      <PackageReference Include="MSTest.TestFramework" Version="3.8.3" />
       <PackageReference Include="System.Text.Json" Version="10.0.0-preview.3.25171.5" />
     </ItemGroup>
